+++ config/permissions.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    [
+        'name' => 'Activity Feed',
+        'flag' => 'activity-feed.index',
+    ],
+    [
+        'name' => 'Delete',
+        'flag' => 'activity-feed.destroy',
+        'parent_flag' => 'activity-feed.index',
+    ],
+];