+++ resources/lang/en/activity-feed.php
@@ -2,4 +2,9 @@
 
 return [
     'name' => 'Activity Feed',
+    'description' => 'Review and manage your site\'s activity feed.',
+    'widget_activity_feed' => 'Activity Feed',
+    'tables' => [
+        'owner_id' => 'Owner',
+    ]
 ];