+++ src/Providers/ActivityFeedServiceProvider.php
@@ -4,7 +4,10 @@ namespace Skillcraft\ActivityFeed\Providers;
 
 use Illuminate\Foundation\AliasLoader;
 use Illuminate\Support\ServiceProvider;
+use Botble\Base\Facades\PanelSectionManager;
+use Botble\Base\PanelSections\PanelSectionItem;
 use Botble\Base\Traits\LoadAndPublishDataTrait;
+use Skillcraft\Core\PanelSections\CorePanelSection;
 use Skillcraft\ActivityFeed\Facades\ActivityFeedHelper;
 
 class ActivityFeedServiceProvider extends ServiceProvider
@@ -18,16 +21,31 @@ class ActivityFeedServiceProvider extends ServiceProvider
     }
 
     public function boot(): void
-    {
-        $this->setNamespace('plugins/activity-feed')
-            ->loadAndPublishConfigurations(['general'])
-            ->loadHelpers()
-            ->loadMigrations()
-            ->loadAndPublishTranslations();
-
+    {    
         if (! is_plugin_active('skillcraft-core')) {
             return;
         }
+
+        $this->setNamespace('plugins/activity-feed')
+            ->loadAndPublishConfigurations(['general', 'permissions'])
+            ->loadHelpers()
+            ->loadMigrations()
+            ->loadAndPublishTranslations()
+            ->loadAndPublishViews()
+            ->loadRoutes();
+            
+            PanelSectionManager::default()->beforeRendering(function () {
+                PanelSectionManager::registerItem(
+                    CorePanelSection::class,
+                    fn () => PanelSectionItem::make('activity-feed')
+                        ->setTitle(trans('plugins/activity-feed::activity-feed.name'))
+                        ->withDescription(trans('plugins/activity-feed::activity-feed.description'))
+                        ->withIcon('ti ti-activity')
+                        ->withPriority(20)
+                        ->withRoute('activity-feed.index')
+                );
+            });
+
 
         $this->app->booted(function () {
             $this->app->register(HookServiceProvider::class);