+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to the Activity Feed plugin will be documented in this file, from latest change to the oldest.
 
+## [1.1.0] - 2024-01-13
+
+### Added
+- [Feature] Added Admin Dashboard Feed Widget 
+- [Feature] Added Resource Table
+- [Enhancement] Resource Permissions
+- [Enhancement] New Activity Feed Model Methods: getOwnerName() & getOwnerLink() & getOwnerAvatarLink()
+
+
 ## [1.0.0] - 2024-01-13
 
 ### Added